Add case handle when account has been disable. Edit get location API. Add admin-api reactivate and deactivate account

+++ src/main/java/com/swp/backend/service/AccountLoginService.java
@@ -39,6 +39,8 @@ public class AccountLoginService {
 
     public void deleteLogin(String userId){
         AccountLoginEntity login = accountLoginRepository.findLoginStateByUserId(userId);
-        accountLoginRepository.deleteById(login.getId());
+        if(login != null){
+            accountLoginRepository.deleteById(login.getId());
+        }
     }
 }
